export thread to upper
